requirements.txt
PyYAML==6.0
six==1.8.0
urllib3==1.26.11
pycg==0.0.6
